Code/requirement.txt
cvxopt==1.3.0
h5py==3.1.0
keras==2.7.0
librosa==0.9.2
matplotlib==3.6.1
numba==0.56.3
numpy==1.19.5
openpyxl==3.0.10
pandas==1.3.0
prettytable==3.4.1
scikit-learn==1.1.2
scipy==1.9.2
spafe==0.2.0
speechpy==2.4
tensorflow-gpu==2.7.0
tqdm==4.63.1
